+++ packages/core/hexhive-data/prisma/migrations/20220531082030_add_org_api_keys/migration.sql
@@ -1,0 +1,12 @@
+-- CreateTable
+CREATE TABLE "APIKey" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "apiKey" TEXT NOT NULL,
+    "organisationId" TEXT NOT NULL,
+
+    CONSTRAINT "APIKey_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "APIKey" ADD CONSTRAINT "APIKey_organisationId_fkey" FOREIGN KEY ("organisationId") REFERENCES "Organisation"("id") ON DELETE CASCADE ON UPDATE CASCADE;