+++ packages/core/hexhive-data/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2-beta.49](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.48...v1.0.2-beta.49) (2022-05-31)
+
+**Note:** Version bump only for package @hexhive/data
+
+
+
+
+
+## [1.0.2-beta.48](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.47...v1.0.2-beta.48) (2022-05-31)
+
+**Note:** Version bump only for package @hexhive/data
+
+
+
+
+
+## [1.0.2-beta.47](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.46...v1.0.2-beta.47) (2022-05-31)
+
+**Note:** Version bump only for package @hexhive/data
+
+
+
+
+
 ## [1.0.2-beta.42](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.41...v1.0.2-beta.42) (2022-05-15)
 
 **Note:** Version bump only for package @hexhive/data